fix bsfilter error message

+++ redcub-server/lib/redcub/deliver.rb
@@ -24,51 +24,64 @@ module RedCub
 
               transaction do
 
-                # debug
-                File.open("/tmp/mail.txt", "wb") do |f|
-                  f.write(mail.data)
+                begin
+                  # debug
+                  File.open("/tmp/mail.txt", "wb") do |f|
+                    f.write(mail.data)
+                  end
+
+                  if mail.data.nil?
+                    mail.destroy
+                    Syslog.info("mail cannot deliverd(Message-ID=#{mail.message_id.gsub(/\%/, '%%')}).")
+                    
+                    next
+                  end
+
+                  tmail = get_tmail_object(mail.data, @myhostname)
+                  
+                  if tmail.from.nil?
+                    tmail.from = mail.mail_from
+                  end
+                  
+                  user = mail.orig_to.split(/@/)[0]
+                  from_id = get_address_id(tmail)
+                  user_id = get_user_id(user)
+                  
+                  filter_id = Filter.filter_id(tmail, user_id)
+                  
+                  header = OrderHash.new
+                  tmail.each_header do |key, value|
+                    header[key] = Base64.decode_b(value.to_s).toutf8
+                  end
+                  
+                  body, content_type = get_message_body(tmail)
+                  Syslog.debug("mail_type: #{content_type}")
+                  
+                  new_mail = Mail.new
+                  new_mail.user_id = user_id
+                  new_mail.message_id = tmail.message_id
+                  new_mail.mail_from_id = from_id
+                  new_mail.filter_id = filter_id
+                  new_mail.receive_date = mail.receive_date
+                  new_mail.subject = tmail.subject.toutf8
+                  new_mail.body_part = get_string_part(body)
+                  new_mail.content_type = content_type
+                  
+                  new_mail.data = tmail.encoded
+                  new_mail.header = header
+                  new_mail.attached_files = get_attached_files(user_id, tmail)
+                  
+                  new_mail.save
+                  
+                  mail.destroy
+
+
+                  Syslog.info("mail deliverd(Message-ID=#{mail.message_id.gsub(/\%/, '%%')}).")
+                rescue
+                  sleep(5)
                 end
-
-                tmail = get_tmail_object(mail.data, @myhostname)
-
-                if tmail.from.nil?
-                  tmail.from = mail.mail_from
-                end
-
-                user = mail.orig_to.split(/@/)[0]
-                from_id = get_address_id(tmail)
-                user_id = get_user_id(user)
-
-                filter_id = Filter.filter_id(tmail, user_id)
-                
-                header = OrderHash.new
-                tmail.each_header do |key, value|
-                  header[key] = Base64.decode_b(value.to_s).toutf8
-                end
-
-                body, content_type = get_message_body(tmail)
-                Syslog.debug("mail_type: #{content_type}")
-                
-                new_mail = Mail.new
-                new_mail.user_id = user_id
-                new_mail.message_id = tmail.message_id
-                new_mail.mail_from_id = from_id
-                new_mail.filter_id = filter_id
-                new_mail.receive_date = mail.receive_date
-                new_mail.subject = tmail.subject.toutf8
-                new_mail.body_part = get_string_part(body)
-                new_mail.content_type = content_type
-
-                new_mail.data = tmail.encoded
-                new_mail.header = header
-                new_mail.attached_files = get_attached_files(user_id, tmail)
-                
-                new_mail.save
-
-                mail.destroy
               end
               
-              Syslog.info("mail deliverd(Message-ID=#{mail.message_id.gsub(/\%/, '%%')}).")
               
             rescue Exception
               write_backtrace
